+++ database/migrations/1762933625806_create_add_uses_publicodes_to_simulateurs_table.ts
@@ -4,14 +4,20 @@ export default class extends BaseSchema {
   protected tableName = 'simulateurs'
 
   async up() {
-    this.schema.alterTable(this.tableName, (table) => {
-      table.boolean('uses_publicodes_forms').defaultTo(false).notNullable()
-    })
+    const hasColumn = await this.schema.hasColumn(this.tableName, 'uses_publicodes_forms')
+    if (!hasColumn) {
+      this.schema.alterTable(this.tableName, (table) => {
+        table.boolean('uses_publicodes_forms').defaultTo(false).notNullable()
+      })
+    }
   }
 
   async down() {
-    this.schema.alterTable(this.tableName, (table) => {
-      table.dropColumn('uses_publicodes_forms')
-    })
+    const hasColumn = await this.schema.hasColumn(this.tableName, 'uses_publicodes_forms')
+    if (hasColumn) {
+      this.schema.alterTable(this.tableName, (table) => {
+        table.dropColumn('uses_publicodes_forms')
+      })
+    }
   }
 }
